Add basic layout and styles (#123)
* Add basic layout and styles.

* Add lower content region and flex grow.

* Update next-drupal.

* Update component naming.

* Test file renaming.

* Test file renaming: add new name.

* Temp rename.

* Rename component files.

* Temp rename layout folder.

* Rename layout folder.

* Move Layout to slug. Change sidebar semantic order. Clean.
