+++ package.json
@@ -17,7 +17,7 @@
     "hds-react": "^1.12.0",
     "html-react-parser": "^1.4.8",
     "next": "12.1.0",
-    "next-drupal": "^1.1.1",
+    "next-drupal": "1.3.0",
     "react": "17.0.2",
     "react-dom": "17.0.2"
   },
