+++ src/components/ContentMapper.tsx
@@ -1,13 +1,13 @@
-import Image from "next/image"
-import Link from "next/link"
+import Image from 'next/image'
+import Link from 'next/link'
 import HtmlBlock from './HtmlBlock'
-import { CONTENT_TYPES, TEXT_HTML_FORMAT } from "@/lib/DRUPAL_API_TYPES"
-import { DrupalParagraph } from "next-drupal";
+import { CONTENT_TYPES, TEXT_HTML_FORMAT } from 'src/lib/drupalApiTypes'
+import { DrupalParagraph } from 'next-drupal';
 
-import { Accordion } from "hds-react"
+import { Accordion } from 'hds-react'
 
 interface ContentMapperProps {
-  content: any;
+  content: any
 }
 
 export function ContentMapper({ content, ...props }: ContentMapperProps): JSX.Element {
