+++ src/components/PreviewAlert.tsx
@@ -1,6 +1,6 @@
-import * as React from "react"
-import { useRouter } from "next/router"
-import Link from "next/link"
+import * as React from 'react'
+import { useRouter } from 'next/router'
+import Link from 'next/link'
 
 export function PreviewAlert() {
   const { isPreview } = useRouter()
