+++ styles/globals.css
@@ -1,9 +1,27 @@
-html,
+:root {
+  --max-width: 1440px;
+}
+
+* {
+  box-sizing: border-box;
+}
+
+html {
+  margin: 0;
+  -webkit-text-size-adjust: 100%;
+  line-height: var(--lineheight-m);
+}
+
 body {
+  background: var(--color-white);
+  color: var(--color-black-90);
+  font-family: var(--font-default);
+  font-size: var(--fontsize-body-m);
+  line-height: var(--lineheight-l);
   padding: 0;
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
-    Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
 }
 
 a {
@@ -11,6 +29,115 @@ a {
   text-decoration: none;
 }
 
-* {
-  box-sizing: border-box;
+ul {
+  margin-top: var(--spacing-l);
+  margin-bottom: var(--spacing-l);
+  padding-top: 0;
+  padding-left: var(--spacing-s);
+  list-style-position: outside;
+}
+
+li {
+  padding-left: var(--spacing-m);
+}
+
+ul li::marker {
+  color: var(--color-metro);
+}
+
+li li::marker {
+  color: var(--color-black-90);
+}
+
+a[class^="link"] {
+  display: flex;
+  text-decoration: none;
+}
+
+a[class^="link"]::after {
+  content: '';
+  margin-left: 10px;
+  width: 1.25rem;
+  height: 1.25rem;
+}
+
+.link-internal::after {
+  background: url('../src/assets/icon-arrow-right.svg') no-repeat center center;
+}
+
+.link-external::after {
+  background: url('../src/assets/icon-link-external.svg') no-repeat center center;
+}
+
+.container {
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.columns {
+  display: flex;
+  flex-direction: column;
+}
+
+.flex-grow {
+  flex-grow: 1;
+}
+
+.flex-order-first {
+  order: -1;
+}
+
+@media (min-width: 768px) {
+  .columns {
+    flex-direction: row;
+  }
+
+  .col {
+    margin-left: var(--spacing-m);
+    margin-right: var(--spacing-m);
+  }
+  
+  .col-3 {
+    width: 25%;
+  }
+  
+  .col-4 {
+    width: 33%;
+  }
+  
+  .col-5 {
+    width: 58%;
+  }
+  
+  .col-6 {
+    width: 50%;
+  }
+  
+  .col-7 {
+    width: 42%;
+  }
+  
+  .col-8 {
+    width: 67%;
+  }
+  
+  .col-9 {
+    width: 75%;
+  }
+
+  .col-12 {
+    width: 100%;
+  }
+}
+
+@media (min-width: 576px) {
+  .col {
+    margin-left: var(--spacing-s);
+    margin-right: var(--spacing-s);
+  }
+}
+
+.col {
+  margin-left: var(--spacing-xs);
+  margin-right: var(--spacing-xs);
 }
