+++ tsconfig.json
@@ -15,9 +15,8 @@
     "jsx": "preserve",
     "baseUrl": "./",
     "paths": {
-      "@/components/*": ["components/*"],
-      "@/nodes/*": ["components/nodes/*"],
-      "@/lib/*": ["lib/*"]
+      "@/components/*": ["src/components/*"],
+      "@/lib/*": ["src/lib/*"]
     },
     "incremental": true
   },
